+++ cmd/neofs-node/config/engine/shard/metabase/metabase.go
@@ -7,15 +7,20 @@ import (
 	"github.com/nspcc-dev/neofs-node/cmd/neofs-node/config/internal"
 )
 
-// PermDefault is the default permission bits for metabase file.
-const PermDefault = 0o640
+const (
+	// PermDefault is the default permission bits for metabase file.
+	PermDefault = 0o640
+	// SearchIterationLimitDefault is the default limit for the number of iterations without finding a match in search.
+	SearchIterationLimitDefault = 10000
+)
 
 // Metabase contains configuration for a metabase.
 type Metabase struct {
-	Path          string        `mapstructure:"path"`
-	Perm          fs.FileMode   `mapstructure:"perm"`
-	MaxBatchSize  internal.Size `mapstructure:"max_batch_size"`
-	MaxBatchDelay time.Duration `mapstructure:"max_batch_delay"`
+	Path                 string        `mapstructure:"path"`
+	Perm                 fs.FileMode   `mapstructure:"perm"`
+	MaxBatchSize         internal.Size `mapstructure:"max_batch_size"`
+	MaxBatchDelay        time.Duration `mapstructure:"max_batch_delay"`
+	SearchIterationLimit *int          `mapstructure:"search_iteration_limit"`
 }
 
 // Normalize metabase configuration by filling in default values.
@@ -30,5 +35,13 @@ func (m *Metabase) Normalize(def Metabase) {
 	m.MaxBatchSize.Check(def.MaxBatchSize, 0)
 	if m.MaxBatchDelay <= 0 && def.MaxBatchDelay > 0 {
 		m.MaxBatchDelay = def.MaxBatchDelay
+	}
+	if m.SearchIterationLimit == nil {
+		if def.SearchIterationLimit != nil {
+			m.SearchIterationLimit = def.SearchIterationLimit
+		} else {
+			v := SearchIterationLimitDefault
+			m.SearchIterationLimit = &v
+		}
 	}
 }